requirements.txt
Pillow
ijson